Add max_connections_per_node on ClusterConnectionPool
This handles a situation where the max_connections argument applies to the
whole cluster, and can starve connections on concurrent processes to nodes
in the cluster. It also results in asymmetrical connection allocation
across the nodes, which makes it challenging to configure file descriptor
and redis max client settings.

The cluster-wide setting can also cause problems if the number
of nodes is greater than max_connections.