fix: ClusterConnectionPool.count_all_num_connections is not thread safe in python 3
In python 3, `dict.values()` is an iterator, whereas in python 2 it
returns a copy of the values as a list.

When running under python 3 and multiple threads, that means the
dictionary _created_connections_per_node is now able to be mutated and
iterated over at the same time, thus raising a RuntimeError.

The race comes from multiple threads calling
`ClusterConnectionPool.make_connection` at the same time.

Coercing the `.values()` to a `list()` _appears_ to the best of my
knowledge to be locked by the GIL and doesn't cause this issue.

I have stress tested this with: https://gist.github.com/mattrobenolt/e88d940c9afa6a18d6b39328eeb9a8f6