fix(client): Possible UnboundLocalError within execute finally block
I'm not sure what all cases trigger this, but within the finally block,
there is no guarantee previously that the variable `connection` was even
defined.

In out case, we hit a SlotNotCoveredError, which cascaded into an
UnboundLocalError from within the finally block.

This change explicitly declares the `connection` variable so it can
safely be checked later if it needs to be freed.